- add tasks per variant for KMP - some smaller code cleanup - add some refined comments
